perf (WORM AI): uses custom shared variables and behavior tree reference to reuse AI code
