+++ Erlang-Legacy/Assets/Core/IA/Behavior/SharedVariable.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: db0bc1e1a0791a642ade33cc1a483555
+guid: 1ef8b3123f75ab1408184301971e5665
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
