+++ Erlang-Legacy/Assets/Core/IA/Behavior/SharedVariable/SharedParticleSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e405ec1f3f5c6a24f98b0c288acc9f93
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
