+++ Erlang-Legacy/Assets/Core/IA/Behavior/SharedVariable/SharedSpriteRenderer.cs.meta
