+++ Erlang-Legacy/Assets/Core/IA/Behavior/Task/Condtional/IsDestroyable.cs
@@ -1,16 +1,11 @@
-﻿using BehaviorDesigner.Runtime;
-using BehaviorDesigner.Runtime.Tasks;
-using UnityEngine;
+﻿using BehaviorDesigner.Runtime.Tasks;
 
 namespace Core.Combat.IA.Conditional
 {
     public class IsDestroyable : EnemyConditional
     {
-        public SharedString hola;
-
         public override TaskStatus OnUpdate()
         {
-            Debug.Log(hola.Value);
             bool destroyed = destroyable.IsDestroyed;
             return destroyed ? TaskStatus.Success : TaskStatus.Failure;
         }
